// src/app/shared/interfaces/user.ts
import { NavigationComponent } from './../../core/navigation/navigation.component';
export interface IUser {

  name?: string;
  id: string;
  tel?: string;
  email: string;
  username?: string;
  password?: string;
  imageUrl?: string;
  // created_at: string;
  // updatedAt: string;
  // __v: number;
}